Fixed Compilation Error
Program actually compiles now (lol). Added a custom login page, added a custom registration page. This app is no longer going to be tinder for dog but instead an app to message people in your local area. Kept the name though and some classes that I'm probably going to remove in future updates.